backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
eventlet==0.36.0
requests==2.31.0
fpdf2==2.7.9
psycopg2-binary==2.9.9
google-generativeai==0.3.2
sendgrid==6.11.0